+++ src/styles/components/checkbox-block.less
@@ -6,8 +6,8 @@
     align-items: center;
   }
 
-  &>div>div:first-of-type {
-    margin-right: 1em;
+  &>div>div:last-of-type {
+    margin-left: auto;
   }
 
   input[type='checkbox'] {
